FROMLIST: refs: use 'uint64_t' for 'ref_update.index'
The 'ref_update.index' variable is used to store an index for a given
reference update. This index is used to order the updates in a
predetermined order, while the default ordering is alphabetical as per
the refname.

For large repositories with millions of references, it should be safer
to use 'uint64_t'. Let's do that. This also is applied for all other
code sections where we store 'index' and pass it around.

Reported-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Karthik Nayak <karthik.188@gmail.com>

Link: https://lore.kernel.org/all/20250122-461-corrupted-reftable-followup-v3-2-ae5f88bf04fa@gmail.com/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>