FROMLIST: fetch set_head: refactor to use remote directly
As a preparatory step to use even more properties from the remote
struct, refactor set_head to take the entire struct as a parameter,
instead of the necessary bits. This also allows consolidating the use of
gtransport->remote in set_head, making the access of the remote's
properties consistent in the function.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>

Link: https://lore.kernel.org/all/20250126220403.289742-1-bence@ferdinandy.com/
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>